+++ Example/Sources/AppDelegate.swift
@@ -29,12 +29,12 @@ final internal class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
 
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil) -> Bool {
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
         window = UIWindow(frame: UIScreen.main.bounds)
-        window?.backgroundColor = .white
-        window?.rootViewController = UINavigationController(rootViewController: InboxViewController())
+        window?.rootViewController = NavigationController(rootViewController: LaunchViewController())
         window?.makeKeyAndVisible()
+        
         return true
     }